--- app/src/main/java/com/gmail/laktionov/lyricsfinder/data/remote/mourits/MouritsLyricsApi.kt
@@ -1,4 +1,0 @@
-package com.gmail.laktionov.lyricsfinder.data.remote.mourits
-
-interface MouritsLyricsApi {
-}
