--- app/src/main/java/com/gmail/laktionov/lyricsfinder/data/remote/mourits/model/MouritsResponse.kt
@@ -1,4 +1,0 @@
-package com.gmail.laktionov.lyricsfinder.data.remote.mourits.model
-
-class MouritsResponse {
-}
