+++ app/src/mourits/java/com/gmail/laktionov/lyricsfinder/di/data/remote/MouritsDataSource.kt
@@ -1,7 +1,7 @@
-package com.gmail.laktionov.lyricsfinder.data.remote.mourits
+package com.gmail.laktionov.lyricsfinder.di.data.remote
 
 import com.gmail.laktionov.lyricsfinder.core.BaseResponse
-import com.gmail.laktionov.lyricsfinder.data.remote.mourits.mapper.MouritsResponseMapper
+import com.gmail.laktionov.lyricsfinder.di.data.remote.mapper.MouritsResponseMapper
 import com.gmail.laktionov.lyricsfinder.domain.RemoteSource
 import com.gmail.laktionov.lyricsfinder.domain.model.SongLyric
