+++ app/src/mourits/java/com/gmail/laktionov/lyricsfinder/di/data/remote/mapper/MouritsResponseMapper.kt
@@ -1,0 +1,4 @@
+package com.gmail.laktionov.lyricsfinder.di.data.remote.mapper
+
+class MouritsResponseMapper {
+}
