+++ app/src/mourits/java/com/gmail/laktionov/lyricsfinder/di/data/remote/model/MouritsRequest.kt
@@ -1,0 +1,4 @@
+package com.gmail.laktionov.lyricsfinder.di.data.remote.model
+
+class MouritsRequest {
+}
